Refactor SAST scan creation logic from FlowService into SastScannrService
